move package demo to org.apache.eventmesh.runtime.demo
