+++ eventmesh-runtime/src/test/java/org/apache/eventmesh/runtime/demo/AsyncSubClient.java
@@ -15,35 +15,32 @@
  * limitations under the License.
  */
 
-package demo;
+package org.apache.eventmesh.runtime.demo;
 
 import io.netty.channel.ChannelHandlerContext;
 
 import org.apache.eventmesh.common.protocol.SubcriptionType;
-import org.apache.eventmesh.common.protocol.tcp.Command;
+import org.apache.eventmesh.common.protocol.tcp.EventMeshMessage;
 import org.apache.eventmesh.common.protocol.tcp.Package;
 
+import client.common.ClientConstants;
 import client.common.MessageUtils;
-import client.common.UserAgentUtils;
 import client.hook.ReceiveMsgHook;
 import client.impl.SubClientImpl;
 import org.apache.eventmesh.common.protocol.SubscriptionMode;
 
-public class CCSubClient {
-
+public class AsyncSubClient {
     public static void main(String[] args) throws Exception {
-        SubClientImpl subClient = new SubClientImpl("127.0.0.1", 10000, UserAgentUtils.createUserAgent());
-        subClient.init();
-        subClient.heartbeat();
-        subClient.listen();
-        subClient.justSubscribe("TEST-TOPIC-TCP-SYNC", SubscriptionMode.CLUSTERING, SubcriptionType.SYNC);
-        subClient.registerBusiHandler(new ReceiveMsgHook() {
+        SubClientImpl client = new SubClientImpl("127.0.0.1", 10002, MessageUtils.generateSubServer());
+        client.init();
+        client.heartbeat();
+        client.justSubscribe(ClientConstants.ASYNC_TOPIC, SubscriptionMode.CLUSTERING, SubcriptionType.ASYNC);
+        client.registerBusiHandler(new ReceiveMsgHook() {
             @Override
             public void handle(Package msg, ChannelHandlerContext ctx) {
-                System.err.println("收到消息: -----------------------------------------" + msg.toString());
-                if (msg.getHeader().getCommand() == Command.REQUEST_TO_CLIENT) {
-                    Package rrResponse = MessageUtils.rrResponse(msg);
-                    ctx.writeAndFlush(rrResponse);
+                if (msg.getBody() instanceof EventMeshMessage) {
+                    String body = ((EventMeshMessage) msg.getBody()).getBody();
+                    System.err.println("receive message -------------------------------" + body);
                 }
             }
         });
